+++ src/gallery.js
@@ -1,0 +1,2 @@
+'use strict';
+
