+++ src/pictures/pictures.js
@@ -2,6 +2,7 @@
 
 var picturesContainer = document.querySelector('.pictures');
 var filtersContainer = document.querySelector('.filters');
+var galleryContainer = document.querySelector('.gallery-overlay');
 var pictures = [];
 var filteredPictures = [];
 
@@ -12,8 +13,47 @@ var pageNumber = 0;
 filtersContainer.classList.add('hidden');
 
 // Отрисовка данных из json файла на основе шаблона
-var renderModule = require('./render');
-var drawPictureElement = renderModule.drawPictureElement;
+// var renderModule = require('./render');
+var drawPictureElement = function(data, container) {
+    var templateElement = document.querySelector('#picture-template');
+
+    if('content' in templateElement) {
+      var element = templateElement.content.children[0].cloneNode(true);
+    } else {
+      element = templateElement.children[0].cloneNode(true);
+    }
+
+    var img = new Image();
+    var elementImg = element.querySelector('img');
+
+    img.onload = function() {
+      elementImg.src = img.src;
+      elementImg.width = 182;
+      elementImg.height = 182;
+    };
+
+    img.onerror = function() {
+      elementImg.classList.add('picture-load-failure');
+    };
+
+    img.src = data.url;
+
+    element.querySelector('.picture-comments').innerHTML = data.comments;
+    element.querySelector('.picture-likes').innerHTML = data.likes;
+
+    getGalleryPictures(data.url);
+
+    element.addEventListener('click', function(event) {
+      event.preventDefault();
+      showGallery(filteredPictures.indexOf(data));
+    });
+
+
+    container.appendChild(element);
+    return element;
+  };
+
+
 
 var renderPictures = function(picturesArray, page, replace) {
   if (replace) {
@@ -59,6 +99,28 @@ var addScroll = function() {
       }
     }, 100);
   });
+};
+
+var galleryPictures = [];
+var indexpic = 0;
+
+var galleryPicture = galleryContainer.querySelector('.gallery-overlay-image');
+
+var activeGalleryPicture = 0;
+
+var getGalleryPictures = function(galleryPicturesArray) {
+  galleryPictures = galleryPicturesArray;
+  console.log(galleryPictures[5]);
+};
+
+var showGallery = function(index) {
+  galleryContainer.classList.remove('invisible');
+
+  indexpic = index;
+}
+
+var setActivePicture = function(index) {
+  galleryPicture.src = galleryPictures[index];
 };
 
 // Фильтрация изображений
